feat: migrate yrb to magnus
This exchanges the underyling library used to create bindings between
Rust and Ruby. The API did not change. Due this being a huge feat, the
commit is not marked as a pure refactoring.

This commit also includes multiple changes to allow for
cross-compilation of gems with the help of rb-sys and rack-compiler.

There are also major improvements to the dev workflow and the Rust
codebase. Including a VSCode setup for clippy.

Co-authored-by: Ian Ker-Seymer <i.kerseymer@gmail.com>