refactor: migrate codebase to yrs 0.13+
This changes y-rb to use yrs 0.13.x+. Even this being a change which
does not bring any features for Ruby users, it unfortunately includes a
few breaking changes. I tried to keep them to a minimum, but yrs behaves
different in certain situations and it cannot be avoided completely.

The best references for what breaking changes you should expect are
the changes made to specs in this PR.